+++ apps/expo-go/modules/@react-native-async-storage/async-storage/lib/module/hooks.js
@@ -1,0 +1,10 @@
+import AsyncStorage from "./AsyncStorage";
+export function useAsyncStorage(key) {
+  return {
+    getItem: (...args) => AsyncStorage.getItem(key, ...args),
+    setItem: (...args) => AsyncStorage.setItem(key, ...args),
+    mergeItem: (...args) => AsyncStorage.mergeItem(key, ...args),
+    removeItem: (...args) => AsyncStorage.removeItem(key, ...args)
+  };
+}
+//# sourceMappingURL=hooks.js.map