+++ apps/expo-go/modules/@react-native-async-storage/async-storage/lib/typescript/shouldFallbackToLegacyNativeModule.d.ts
@@ -1,0 +1,2 @@
+export declare function shouldFallbackToLegacyNativeModule(): boolean;
+//# sourceMappingURL=shouldFallbackToLegacyNativeModule.d.ts.map