+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -35,6 +35,8 @@ import { PlatformBundlers } from '../platformBundlers';
 
 type Mutable<T> = { -readonly [K in keyof T]: T[K] };
 
+const ASSET_REGISTRY_SRC = `const assets=[];module.exports={registerAsset:s=>assets.push(s),getAssetByID:s=>assets[s-1]};`;
+
 const debug = require('debug')('expo:start:server:metro:multi-platform') as typeof console.log;
 
 function withWebPolyfills(
@@ -132,6 +134,7 @@ export function withExtendedResolver(
     isFastResolverEnabled,
     isExporting,
     isReactCanaryEnabled,
+    isReactServerComponentsEnabled,
     getMetroBundler,
   }: {
     tsconfig: TsConfigPaths | null;
@@ -139,46 +142,21 @@ export function withExtendedResolver(
     isFastResolverEnabled?: boolean;
     isExporting?: boolean;
     isReactCanaryEnabled?: boolean;
+    isReactServerComponentsEnabled?: boolean;
     getMetroBundler: () => Bundler;
   }
 ) {
+  if (isReactServerComponentsEnabled) {
+    Log.warn(
+      `Experimental React Server Components is enabled. Production exports are not supported yet.`
+    );
+  }
   if (isFastResolverEnabled) {
     Log.warn(`Experimental module resolution is enabled.`);
   }
 
   if (isReactCanaryEnabled) {
     Log.warn(`Experimental React Canary version is enabled.`);
-  }
-
-  let _assetRegistryPath: string | null = null;
-
-  // Fetch this lazily for testing purposes.
-  function getAssetRegistryPath() {
-    if (_assetRegistryPath) {
-      return _assetRegistryPath;
-    }
-
-    // Get the `transformer.assetRegistryPath`
-    // this needs to be unified since you can't dynamically
-    // swap out the transformer based on platform.
-    if (
-      config.transformer.assetRegistryPath &&
-      path.isAbsolute(config.transformer.assetRegistryPath)
-    ) {
-      _assetRegistryPath = fs.realpathSync(config.transformer.assetRegistryPath);
-      return _assetRegistryPath;
-    }
-
-    const assetRegistryPath = fs.realpathSync(
-      path.resolve(
-        resolveFrom(
-          config.projectRoot,
-          config.transformer.assetRegistryPath ?? '@react-native/assets-registry/registry.js'
-        )
-      )
-    );
-    _assetRegistryPath = assetRegistryPath;
-    return assetRegistryPath;
   }
 
   const defaultResolver = metroResolver.resolve;
@@ -214,6 +192,14 @@ export function withExtendedResolver(
     if (resolveFrom.silent(config.projectRoot, '@expo/vector-icons')) {
       debug('Enabling alias: react-native-vector-icons -> @expo/vector-icons');
       _universalAliases.push([/^react-native-vector-icons(\/.*)?/, '@expo/vector-icons$1']);
+    }
+    if (isReactServerComponentsEnabled) {
+      if (resolveFrom.silent(config.projectRoot, 'expo-router/rsc')) {
+        debug('Enabling bridge alias: expo-router -> expo-router/rsc');
+        _universalAliases.push([/^expo-router$/, 'expo-router/rsc']);
+        // Bridge the internal entry point which is a standalone import to ensure package.json resolution works as expected.
+        _universalAliases.push([/^expo-router\/entry-classic$/, 'expo-router/rsc/entry']);
+      }
     }
     return _universalAliases;
   }
@@ -303,6 +289,18 @@ export function withExtendedResolver(
   const idFactory =
     config.serializer?.createModuleIdFactory?.() ?? ((id: number | string): number | string => id);
 
+  const getAssetRegistryModule = () => {
+    const virtualModuleId = `\0polyfill:assets-registry`;
+    getMetroBundlerWithVirtualModules(getMetroBundler()).setVirtualModule(
+      virtualModuleId,
+      ASSET_REGISTRY_SRC
+    );
+    return {
+      type: 'sourceFile',
+      filePath: virtualModuleId,
+    } as const;
+  };
+
   // If Node.js pass-through, then remap to a module like `module.exports = $$require_external(<module>)`.
   // If module should be shimmed, remap to an empty module.
   const externals: {
@@ -355,7 +353,7 @@ export function withExtendedResolver(
         }
 
         const isExternal = // Extern these modules in standard Node.js environments.
-          /^(styleq(\/.+)?|deprecated-react-native-prop-types|react-native-safe-area-context|invariant|nullthrows|memoize-one|@react-native\/assets-registry\/registry|react|react\/jsx-dev-runtime|scheduler|expo-modules-core|react-native|react-dom(\/.+)?|metro-runtime(\/.+)?)$/.test(
+          /^(styleq(\/.+)?|deprecated-react-native-prop-types|react-native-safe-area-context|invariant|nullthrows|memoize-one|react|react\/jsx-dev-runtime|scheduler|expo-modules-core|react-native|react-dom(\/.+)?|metro-runtime(\/.+)?)$/.test(
             moduleName
           ) ||
           /^react-native-web\/dist\/exports\/(Platform|NativeEventEmitter|StyleSheet|NativeModules|DeviceEventEmitter|Text|View)$/.test(
@@ -542,21 +540,27 @@ export function withExtendedResolver(
       return null;
     },
 
-    // TODO: Reduce these as much as possible in the future.
-    // Complex post-resolution rewrites.
+    // Polyfill for asset registry
     (context: ResolutionContext, moduleName: string, platform: string | null) => {
-      const doResolve = getStrictResolver(context, platform);
+      if (/^@react-native\/assets-registry\/registry(\.js)?$/.test(moduleName)) {
+        return getAssetRegistryModule();
+      }
 
       if (
         platform === 'web' &&
         context.originModulePath.match(/node_modules[\\/]react-native-web[\\/]/) &&
         moduleName.includes('/modules/AssetRegistry')
       ) {
-        return {
-          type: 'sourceFile',
-          filePath: getAssetRegistryPath(),
-        };
+        return getAssetRegistryModule();
       }
+
+      return null;
+    },
+
+    // TODO: Reduce these as much as possible in the future.
+    // Complex post-resolution rewrites.
+    (context: ResolutionContext, moduleName: string, platform: string | null) => {
+      const doResolve = getStrictResolver(context, platform);
 
       const result = doResolve(moduleName);
 
@@ -668,13 +672,29 @@ export function withExtendedResolver(
         context.unstable_enablePackageExports = true;
         context.unstable_conditionsByPlatform = {};
 
-        if (platform === 'web') {
-          // Node.js runtimes should only be importing main at the moment.
-          // This is a temporary fix until we can support the package.json exports.
-          context.mainFields = ['main', 'module'];
+        const isReactServerComponents =
+          context.customResolverOptions?.environment === 'react-server';
+
+        if (isReactServerComponents) {
+          // NOTE: Align the behavior across server and client. This is a breaking change so we'll just roll it out with React Server Components.
+          // This ensures that react-server and client code both resolve `module` and `main` in the same order.
+          if (platform === 'web') {
+            // Node.js runtimes should only be importing main at the moment.
+            // This is a temporary fix until we can support the package.json exports.
+            context.mainFields = ['module', 'main'];
+          } else {
+            // In Node.js + native, use the standard main fields.
+            context.mainFields = ['react-native', 'module', 'main'];
+          }
         } else {
-          // In Node.js + native, use the standard main fields.
-          context.mainFields = ['react-native', 'main', 'module'];
+          if (platform === 'web') {
+            // Node.js runtimes should only be importing main at the moment.
+            // This is a temporary fix until we can support the package.json exports.
+            context.mainFields = ['main', 'module'];
+          } else {
+            // In Node.js + native, use the standard main fields.
+            context.mainFields = ['react-native', 'main', 'module'];
+          }
         }
 
         // Enable react-server import conditions.
@@ -722,21 +742,21 @@ export async function withMetroMultiPlatformAsync(
     exp,
     platformBundlers,
     isTsconfigPathsEnabled,
-    webOutput,
     isFastResolverEnabled,
     isExporting,
     isReactCanaryEnabled,
     isNamedRequiresEnabled,
+    isReactServerComponentsEnabled,
     getMetroBundler,
   }: {
     config: ConfigT;
     exp: ExpoConfig;
     isTsconfigPathsEnabled: boolean;
     platformBundlers: PlatformBundlers;
-    webOutput?: 'single' | 'static' | 'server';
     isFastResolverEnabled?: boolean;
     isExporting?: boolean;
     isReactCanaryEnabled: boolean;
+    isReactServerComponentsEnabled: boolean;
     isNamedRequiresEnabled: boolean;
     getMetroBundler: () => Bundler;
   }
@@ -802,6 +822,7 @@ export async function withMetroMultiPlatformAsync(
     isTsconfigPathsEnabled,
     isFastResolverEnabled,
     isReactCanaryEnabled,
+    isReactServerComponentsEnabled,
     getMetroBundler,
   });
 }