+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -536,6 +536,7 @@ class ExpoCameraView(
   }
 
   fun releaseCamera() = appContext.mainQueue.launch {
+    shouldCreateCamera = true
     cameraProvider?.unbindAll()
   }