+++ packages/expo-router/src/fork/createMemoryHistory.ts
@@ -15,7 +15,7 @@ type HistoryRecord = {
   path: string;
 };
 
-export default function createMemoryHistory() {
+export function createMemoryHistory() {
   let index = 0;
   let items: HistoryRecord[] = [];