+++ packages/expo-updates/ios/EXUpdates/Events/UpdatesEventManager.swift
@@ -3,12 +3,6 @@
 import ExpoModulesCore
 
 public protocol UpdatesEventManager: AnyObject {
-  /**
-   The AppContext from expo-modules-core.
-   This is optional, but required for expo-updates module events to work.
-   */
-  var appContext: AppContext? { get set }
-  var shouldEmitJsEvents: Bool { get set }
-
-  func sendUpdateStateChangeEventToAppContext(_ eventType: UpdatesStateEventType, context: UpdatesStateContext)
+  var eventEmitter: EXEventEmitterService? { get set }
+  func sendStateMachineContextEvent(context: UpdatesStateContext)
 }